+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+# 0.9.46 (2021-08-06)
+* [feat][image] 提供preserveAspectRatio配置
+* [feat] 富文本`RichText`移植
+* [feat] 新增`CustomElement`
+* [path] path2d.connect method
 # 0.9.45(2021-07-23)
 * feat: 提供set get rAF接口